feat: add minidraco/three/single single-threaded entry point
Apps that only ever decode on the main thread (workers: false) still paid
for the worker chunk: the `new Worker(new URL('./worker.js', import.meta.url))`
expression in three/index.ts makes bundlers emit worker.js — a second full
copy of the decoder (~22 KB brotli) — even when it is never used.

Split the DRACOLoader into a single-threaded core (three/loader-base.ts, zero
worker references) that the new `minidraco/three/single` entry re-exports, and
a worker-pool subclass (three/index.ts) that adds the pool. Only the pool
module contains the `new Worker(...)` expression, so importing the single
entry never triggers the worker chunk.

`minidraco/three` is unchanged — same MiniDRACOLoader, same `workers` boolean
for mixed use (some loaders pooled, some not).

- dist/three-single.js: 0 worker references; dist/three.js: unchanged (2).
- Same public API (MiniDRACOLoader / DRACOLoader, decodeDracoFile, ...).
- 49 tests pass; example still builds against minidraco/three.
